requirements.txt
requests
pyzbar
qrcode
Pillow
pyinstaller
